# api-doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coco Travel API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tours": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Get Tours",
        "operationId": "get_tours_api_tours_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Featured"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourResponse"
                  },
                  "title": "Response Get Tours Api Tours Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tours"
        ],
        "summary": "Create Tour",
        "operationId": "create_tour_api_tours_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tours/search": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Search Tours",
        "operationId": "search_tours_api_tours_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tours/category/{category_slug}": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Get Tours By Category",
        "operationId": "get_tours_by_category_api_tours_category__category_slug__get",
        "parameters": [
          {
            "name": "category_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tours/{tour_id}": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Get Tour",
        "operationId": "get_tour_api_tours__tour_id__get",
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tour Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tours"
        ],
        "summary": "Update Tour",
        "operationId": "update_tour_api_tours__tour_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tour Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tours"
        ],
        "summary": "Delete Tour",
        "operationId": "delete_tour_api_tours__tour_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tour Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TourCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "duration",
          "location",
          "images",
          "category_id"
        ],
        "title": "TourCreate"
      },
      "TourResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "duration",
          "location",
          "images",
          "category_id",
          "id",
          "slug",
          "created_at",
          "updated_at",
          "is_active"
        ],
        "title": "TourResponse"
      },
      "TourUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "featured": {
            "type": "boolean",
            "title": "Featured",
            "default": false
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "duration",
          "location",
          "images",
          "category_id"
        ],
        "title": "TourUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "api/auth/token"
          }
        }
      }
    }
  }
}